Organize certificate component files into directories.
