+++ apps/admin-portal/src/components/certificates/import-certificate.tsx
@@ -23,8 +23,7 @@ import * as forge from "node-forge";
 import React, { FunctionComponent, ReactElement, useEffect, useState } from "react";
 import { useDispatch } from "react-redux";
 import { Grid, Icon, Modal } from "semantic-ui-react";
-import { CertificateSummary } from "./certificate-summary";
-import { UploadCertificate } from "./upload-certificate";
+import { CertificateSummary, UploadCertificate } from "./wizard";
 import { createKeystoreCertificate } from "../../api";
 import { ApplicationWizardStepIcons } from "../../configs";
 import { AlertLevels, Certificate, DisplayCertificate } from "../../models";
